requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
